+++ docs/changelog.md
@@ -8,6 +8,12 @@
 
 See the [GitHub "Features" project](https://github.com/users/peter88213/projects/1).
 
+### v4.33.0
+
+- Tree browsing history implemented.
+
+Based on PyWriter v12.13.6
+
 ### v4.32.0
 
 - Add a "Go to associated scene" to the "Arc point" properties.